prompt_engineering/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0

# For Technique 6 (structured JSON → PostgreSQL) in 01_prompt_techniques.py
psycopg2-binary>=2.9.0
